+++ Editor/SafeAreaEvaluationModeDrawer.cs
@@ -1,0 +1,7 @@
+using UnityEditor;
+
+namespace E7.NotchSolution
+{
+    [CustomPropertyDrawer(typeof(SafeAreaEvaluationMode))]
+    public class SafeAreaEvaluationModeDrawer : EnumButtonsDrawer { }
+}