+++ Editor/SafeAreaPaddingDrawer.cs
@@ -13,26 +13,7 @@ namespace E7.NotchSolution
             var landscape = serializedObject.FindProperty("landscapePaddings");
             var influence = serializedObject.FindProperty("influence");
 
-            bool landscapeCompatible = false;
-            bool portraitCompatible = false;
-
-            switch (PlayerSettings.defaultInterfaceOrientation)
-            {
-                case UIOrientation.LandscapeLeft:
-                case UIOrientation.LandscapeRight:
-                    landscapeCompatible = true;
-                    break;
-                case UIOrientation.Portrait:
-                case UIOrientation.PortraitUpsideDown:
-                    portraitCompatible = true;
-                    break;
-                case UIOrientation.AutoRotation:
-                    if (PlayerSettings.allowedAutorotateToLandscapeLeft) landscapeCompatible = true;
-                    if (PlayerSettings.allowedAutorotateToLandscapeRight) landscapeCompatible = true;
-                    if (PlayerSettings.allowedAutorotateToPortrait) portraitCompatible = true;
-                    if (PlayerSettings.allowedAutorotateToPortraitUpsideDown) portraitCompatible = true;
-                    break;
-            }
+            (bool landscapeCompatible, bool portraitCompatible) = NotchSolutionUtility.GetOrientationCompatibility();
 
             if (portraitCompatible && landscapeCompatible)
             {
@@ -40,7 +21,7 @@ namespace E7.NotchSolution
                 EditorGUILayout.Separator();
             }
 
-            bool dual = orientationType.enumValueIndex == (int)SafeAreaPaddingOrientationType.DualOrientation;
+            bool dual = orientationType.enumValueIndex == (int)SupportedOrientations.Dual;
 
             if (dual)
             {