+++ Editor/SafeAreaPaddingOrientationTypeDrawer.cs
@@ -1,0 +1,9 @@
+﻿using UnityEditor;
+using UnityEngine;
+using System.Linq;
+
+namespace E7.NotchSolution
+{
+    [CustomPropertyDrawer(typeof(SupportedOrientations))]
+    public class SupportedOrientationsDrawer : EnumButtonsDrawer { }
+}