+++ Runtime/INotchSimulatorTarget.cs
@@ -4,6 +4,6 @@ namespace E7.NotchSolution
 {
     public interface INotchSimulatorTarget
     {
-        void SimulatorUpdate(Rect simulatedSafeArea, Rect[] simulatedCutouts);
+        void SimulatorUpdate(Rect simulatedSafeAreaRelative, Rect[] simulatedCutoutsRelative);
     }
 }