+++ Runtime/PerEdgeEvaluationModes.cs
@@ -1,0 +1,7 @@
+using System;
+
+namespace E7.NotchSolution
+{
+    [Serializable]
+    public class PerEdgeEvaluationModes : PerEdgeValues<SafeAreaEvaluationMode> { }
+}