+++ Runtime/SafeAreaAdaptation/PerEdgeAdaptations.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+namespace E7.NotchSolution
+{
+    [Serializable]
+    public class PerEdgeAdaptations : PerEdgeValues<AdaptationByAnimationClips> { }
+}