+++ Runtime/SafeAreaPadding/SafeAreaPaddingOrientationType.cs
@@ -1,8 +1,0 @@
-namespace E7.NotchSolution
-{
-    public enum SafeAreaPaddingOrientationType
-    {
-        SingleOrientation,
-        DualOrientation,
-    }
-}